refactor(default-layout): migrate `CreateDocumentDialog` (previous name `ActionDialog`) to `@sanity/ui`
