+++ packages/@sanity/default-layout/src/createDocumentDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateDocumentDialog'
